\name{homecare}
\docType{data}
\alias{homecare}
\title{Occupation and Home Care Eligibility}
\description{
  Occupation and home care eligibility for 383
  patients medically elgibile for home care.
}
\usage{homecare}
\format{A data frame containing two variables, \code{occupation} and \code{eligibility}, over 383 observations.}
\source{Dunn, O. J. (1964). \dQuote{Multiple comparisons using rank sums.} \emph{Technometrics}, 6(3):241–252s. Page 244, Table 1}
\references{
  The data have been adapted from a study of a group of patients entering the Los Angeles County General Hospital during the years 1959-61. 
}
\keyword{datasets}
